backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
aiohttp==3.9.3
python-multipart==0.0.9
google-generativeai==0.7.0
tavily-python>=0.3.3 